requirements.txt
# Core ML
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.24.0

# Auto-labeling (package discovery)
autodistill>=0.1.0
autodistill-grounding-dino>=0.1.0

# Google Vision fallback
google-cloud-vision>=3.5.0

# Frigate API & HTTP
requests>=2.31.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0

# Label Studio integration
label-studio-sdk>=1.0.0

# CLI & config
click>=8.1.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Scheduling
schedule>=1.2.0

# ONNX export validation
onnx>=1.14.0
onnxruntime>=1.16.0
